Modify cps result float to integer

+++ progamertest/cps/forms.py
@@ -1,4 +1,4 @@
 from django import forms
 from .models import CpsResult
 class CpsForm(forms.Form):
-    cps_result = forms.FloatField(widget=forms.HiddenInput())
+    cps_result = forms.IntegerField(widget=forms.HiddenInput())
